Add proc macros to generate Service/Client traits
RpcService generates a Service trait, and RpcClient for a Client trait.
Only Enum can derive the macro, each enum variant represents a possible
RPC method, and the variant values serve as parameters for these methods.
There's an optional attribute `rpc_service(serialize = "bincode" | "serde")`
to derive the appropriate serialization traits for the Request and
Response structs. Currently, only bincode is implemented, but this
should be a feature flag.